add logic and SwapButton to handle change network.
